fastapi_ai/requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
faiss-cpu
numpy
pydantic
python-dotenv
